+++ webview/src/setup/components/App.test.tsx
@@ -31,12 +31,14 @@ const DEFAULT_DATA = {
     version: '1.0.0'
   },
   hasData: false,
+  isAboveLatestTestedVersion: false,
   isPythonExtensionUsed: false,
   isStudioConnected: false,
   needsGitCommit: false,
   needsGitInitialized: false,
   projectInitialized: true,
   pythonBinPath: undefined,
+  remoteList: undefined,
   sectionCollapsed: undefined,
   shareLiveToStudio: false
 }
@@ -95,6 +97,9 @@ describe('App', () => {
       })
 
       expect(screen.getByText('DVC is incompatible')).toBeInTheDocument()
+      expect(
+        screen.getByText('Please update your install and try again.')
+      ).toBeInTheDocument()
 
       const button = screen.getByText('Check Compatibility')
       expect(button).toBeInTheDocument()
@@ -105,18 +110,27 @@ describe('App', () => {
       })
     })
 
-    it('should show a screen saying that DVC is not installed if the cli is unavailable', () => {
+    it('should tell the user than they can auto upgrade DVC if DVC is incompatible and python is available', () => {
       renderApp({
-        cliCompatible: undefined,
+        cliCompatible: false,
         dvcCliDetails: {
           command: 'dvc',
-          version: undefined
-        }
+          version: '1.0.0'
+        },
+        pythonBinPath: 'python'
       })
 
-      expect(screen.getAllByText('DVC is currently unavailable')).toHaveLength(
-        2
-      )
+      expect(screen.getByText('DVC is incompatible')).toBeInTheDocument()
+
+      const compatibityButton = screen.getByText('Check Compatibility')
+      expect(compatibityButton).toBeInTheDocument()
+      const upgradeButton = screen.getByText('Upgrade (pip)')
+      expect(upgradeButton).toBeInTheDocument()
+
+      fireEvent.click(upgradeButton)
+      expect(mockPostMessage).toHaveBeenCalledWith({
+        type: MessageFromWebviewType.UPGRADE_DVC
+      })
     })
 
     it('should tell the user they cannot install DVC without a Python interpreter', () => {
@@ -221,7 +235,12 @@ describe('App', () => {
     it('should show a screen saying that DVC is not initialized if the project is not initialized and git is uninitialized', () => {
       renderApp({ needsGitInitialized: true, projectInitialized: false })
 
-      expect(screen.getByText('DVC is not initialized')).toBeInTheDocument()
+      const uninitializedText = screen.getAllByText('DVC is not initialized')
+
+      expect(uninitializedText).toHaveLength(2)
+      for (const text of uninitializedText) {
+        expect(text).toBeInTheDocument()
+      }
     })
 
     it('should offer to initialize Git if it is possible', () => {
@@ -244,19 +263,13 @@ describe('App', () => {
         projectInitialized: false
       })
 
-      expect(screen.queryByText('Initialize Git')).not.toBeInTheDocument()
-    })
+      const uninitialized = screen.getAllByText('DVC is not initialized')
 
-    it('should show a screen saying that DVC is not initialized if the project is not initialized and dvc is installed', () => {
-      renderApp({
-        projectInitialized: false
-      })
-
-      expect(screen.getByText('DVC is not initialized')).toBeInTheDocument()
+      expect(uninitialized).toHaveLength(4)
     })
 
     it('should not show a screen saying that DVC is not initialized if the project is initialized and dvc is installed', () => {
-      renderApp()
+      renderApp({ remoteList: { mocRoot: undefined } })
 
       expect(
         screen.queryByText('DVC is not initialized')
@@ -281,8 +294,9 @@ describe('App', () => {
         projectInitialized: false,
         sectionCollapsed: {
           [SetupSection.DVC]: false,
-          [SetupSection.STUDIO]: true,
-          [SetupSection.EXPERIMENTS]: true
+          [SetupSection.EXPERIMENTS]: true,
+          [SetupSection.REMOTES]: true,
+          [SetupSection.STUDIO]: true
         }
       }
 
@@ -393,7 +407,9 @@ describe('App', () => {
         projectInitialized: false
       })
 
-      const iconWrapper = screen.getAllByTestId('info-tooltip-toggle')[0]
+      const iconWrapper = within(
+        screen.getByTestId('dvc-section-details')
+      ).getByTestId('info-tooltip-toggle')
 
       expect(
         within(iconWrapper).getByTestId(TooltipIconType.ERROR)
@@ -401,12 +417,37 @@ describe('App', () => {
     })
 
     it('should show a passed icon if DVC CLI is compatible and project is initialized', () => {
-      renderApp()
+      renderApp({ remoteList: { mockRoot: undefined } })
+      expect(screen.queryByText('DVC is not setup')).not.toBeInTheDocument()
 
-      const iconWrapper = screen.getAllByTestId('info-tooltip-toggle')[0]
+      const iconWrapper = within(
+        screen.getByTestId('dvc-section-details')
+      ).getByTestId('info-tooltip-toggle')
 
       expect(
         within(iconWrapper).getByTestId(TooltipIconType.PASSED)
+      ).toBeInTheDocument()
+    })
+
+    it('should add a warning icon and message if version is above the latest tested version', () => {
+      renderApp({
+        isAboveLatestTestedVersion: true
+      })
+
+      const iconWrapper = within(
+        screen.getByTestId('dvc-section-details')
+      ).getByTestId('info-tooltip-toggle')
+
+      expect(
+        within(iconWrapper).getByTestId(TooltipIconType.WARNING)
+      ).toBeInTheDocument()
+
+      fireEvent.mouseEnter(iconWrapper)
+
+      expect(
+        screen.getByText(
+          'The located version has not been tested against the extension. If you are experiencing unexpected behaviour, first try to update the extension. If there are no updates available, please downgrade DVC to the same minor version as displayed and reload the window.'
+        )
       ).toBeInTheDocument()
     })
   })
@@ -422,7 +463,8 @@ describe('App', () => {
 
     it('should open the dvc section when clicking the Setup DVC button on the dvc is not setup screen', () => {
       renderApp({
-        projectInitialized: false
+        projectInitialized: false,
+        remoteList: { mockRoot: undefined }
       })
 
       const experimentsText = screen.getByText('DVC is not setup')
@@ -496,7 +538,9 @@ describe('App', () => {
     it('should show an error icon if experiments are not setup', () => {
       renderApp()
 
-      const iconWrapper = screen.getAllByTestId('info-tooltip-toggle')[1]
+      const iconWrapper = within(
+        screen.getByTestId('experiments-section-details')
+      ).getByTestId('info-tooltip-toggle')
 
       expect(
         within(iconWrapper).getByTestId(TooltipIconType.ERROR)
@@ -508,7 +552,9 @@ describe('App', () => {
         cliCompatible: false
       })
 
-      const iconWrapper = screen.getAllByTestId('info-tooltip-toggle')[1]
+      const iconWrapper = within(
+        screen.getByTestId('experiments-section-details')
+      ).getByTestId('info-tooltip-toggle')
 
       expect(
         within(iconWrapper).getByTestId(TooltipIconType.ERROR)
@@ -520,7 +566,9 @@ describe('App', () => {
         hasData: true
       })
 
-      const iconWrapper = screen.getAllByTestId('info-tooltip-toggle')[1]
+      const iconWrapper = within(
+        screen.getByTestId('experiments-section-details')
+      ).getByTestId('info-tooltip-toggle')
 
       expect(
         within(iconWrapper).getByTestId(TooltipIconType.PASSED)
@@ -578,7 +626,9 @@ describe('App', () => {
         cliCompatible: false
       })
 
-      const iconWrapper = screen.getAllByTestId('info-tooltip-toggle')[2]
+      const iconWrapper = within(
+        screen.getByTestId('studio-section-details')
+      ).getByTestId('info-tooltip-toggle')
 
       expect(
         within(iconWrapper).getByTestId(TooltipIconType.ERROR)
@@ -588,7 +638,9 @@ describe('App', () => {
     it('should show an info icon if dvc is compatible but studio is not connected', () => {
       renderApp()
 
-      const iconWrapper = screen.getAllByTestId('info-tooltip-toggle')[2]
+      const iconWrapper = within(
+        screen.getByTestId('studio-section-details')
+      ).getByTestId('info-tooltip-toggle')
 
       expect(
         within(iconWrapper).getByTestId(TooltipIconType.INFO)
@@ -629,7 +681,9 @@ describe('App', () => {
         isStudioConnected: true
       })
 
-      const iconWrapper = screen.getAllByTestId('info-tooltip-toggle')[2]
+      const iconWrapper = within(
+        screen.getByTestId('studio-section-details')
+      ).getByTestId('info-tooltip-toggle')
 
       expect(
         within(iconWrapper).getByTestId(TooltipIconType.PASSED)
@@ -646,9 +700,10 @@ describe('App', () => {
       renderApp({
         isStudioConnected: true,
         sectionCollapsed: {
+          [SetupSection.DVC]: false,
           [SetupSection.EXPERIMENTS]: true,
-          [SetupSection.STUDIO]: true,
-          [SetupSection.DVC]: false
+          [SetupSection.REMOTES]: true,
+          [SetupSection.STUDIO]: true
         }
       })
       mockPostMessage.mockClear()
@@ -667,9 +722,10 @@ describe('App', () => {
       renderApp({
         isStudioConnected: true,
         sectionCollapsed: {
+          [SetupSection.DVC]: true,
           [SetupSection.EXPERIMENTS]: false,
-          [SetupSection.STUDIO]: true,
-          [SetupSection.DVC]: true
+          [SetupSection.REMOTES]: true,
+          [SetupSection.STUDIO]: true
         }
       })
       mockPostMessage.mockClear()
@@ -688,9 +744,10 @@ describe('App', () => {
       renderApp({
         isStudioConnected: true,
         sectionCollapsed: {
+          [SetupSection.DVC]: true,
           [SetupSection.EXPERIMENTS]: true,
-          [SetupSection.STUDIO]: false,
-          [SetupSection.DVC]: true
+          [SetupSection.REMOTES]: true,
+          [SetupSection.STUDIO]: false
         }
       })
       mockPostMessage.mockClear()
@@ -703,6 +760,97 @@ describe('App', () => {
       const experiments = screen.getByText('Experiments')
       expect(experiments).toBeVisible()
       expect(screen.getByText(experimentsText)).not.toBeVisible()
+    })
+  })
+
+  describe('Remotes', () => {
+    it('should show the setup DVC button if the remoteList is undefined (no projects)', () => {
+      renderApp({
+        remoteList: undefined,
+        sectionCollapsed: {
+          [SetupSection.DVC]: true,
+          [SetupSection.EXPERIMENTS]: true,
+          [SetupSection.REMOTES]: false,
+          [SetupSection.STUDIO]: true
+        }
+      })
+
+      const setupDVCButton = screen.getByText('Setup DVC')
+
+      expect(setupDVCButton).toBeInTheDocument()
+      expect(setupDVCButton).toBeVisible()
+    })
+
+    it('should show the connect to remote storage screen if no remotes are connected', () => {
+      renderApp({
+        remoteList: { demo: undefined, 'example-get-started': undefined },
+        sectionCollapsed: {
+          [SetupSection.DVC]: true,
+          [SetupSection.EXPERIMENTS]: true,
+          [SetupSection.REMOTES]: false,
+          [SetupSection.STUDIO]: true
+        }
+      })
+
+      const setupDVCButton = screen.getByText('Connect to Remote Storage')
+
+      expect(setupDVCButton).toBeInTheDocument()
+      expect(setupDVCButton).toBeVisible()
+    })
+
+    it('should show the list of remotes if there is only one project in the workspace', () => {
+      renderApp({
+        remoteList: {
+          'example-get-started': { drive: 'gdrive://appDataFolder' }
+        },
+        sectionCollapsed: {
+          [SetupSection.DVC]: true,
+          [SetupSection.EXPERIMENTS]: true,
+          [SetupSection.REMOTES]: false,
+          [SetupSection.STUDIO]: true
+        }
+      })
+
+      const setupDVCButton = screen.getByText('Remote Storage Connected')
+
+      expect(setupDVCButton).toBeInTheDocument()
+      expect(setupDVCButton).toBeVisible()
+
+      expect(screen.getByText('drive')).toBeInTheDocument()
+      expect(screen.getByText('gdrive://appDataFolder')).toBeInTheDocument()
+      expect(screen.queryByText('example-get-started')).not.toBeInTheDocument()
+    })
+
+    it('should show the list of remotes by project if there are multiple projects and one has remote(s) connected', () => {
+      renderApp({
+        remoteList: {
+          demo: undefined,
+          'example-get-started': {
+            drive: 'gdrive://appDataFolder',
+            storage: 's3://some-bucket'
+          }
+        },
+        sectionCollapsed: {
+          [SetupSection.DVC]: true,
+          [SetupSection.EXPERIMENTS]: true,
+          [SetupSection.REMOTES]: false,
+          [SetupSection.STUDIO]: true
+        }
+      })
+
+      const setupDVCButton = screen.getByText('Remote Storage Connected')
+
+      expect(setupDVCButton).toBeInTheDocument()
+      expect(setupDVCButton).toBeVisible()
+
+      expect(screen.getByText('demo')).toBeInTheDocument()
+      expect(screen.getAllByText('-')).toHaveLength(2)
+
+      expect(screen.getByText('example-get-started')).toBeInTheDocument()
+      expect(screen.getByText('drive')).toBeInTheDocument()
+      expect(screen.getByText('gdrive://appDataFolder')).toBeInTheDocument()
+      expect(screen.getByText('storage')).toBeInTheDocument()
+      expect(screen.getByText('s3://some-bucket')).toBeInTheDocument()
     })
   })
 })